+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "blocke-cli-api",
-  "version": "0.25.1",
+  "version": "0.25.2",
   "author": "edinaldofox <edinaldofox@gmail.com>",
   "description": "Command-line tool to query the blockchain for a variety of cryptos. View blocks, transactions, and accounts (where applicable).",
   "keywords": [
